+++ src/app/src/components/FilterSidebar.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { bool, func, oneOf } from 'prop-types';
 import { connect } from 'react-redux';
+import { withStyles } from '@material-ui/core';
 import AppBar from '@material-ui/core/AppBar';
 import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
@@ -42,12 +43,19 @@ import {
 
 import { allListsAreEmpty } from '../util/util';
 
-const filterSidebarStyles = Object.freeze({
-    controlPanelStyles: Object.freeze({
-        height: 'inherit',
-        width: 'inherit',
-    }),
+const controlPanelStyles = Object.freeze({
+    height: 'inherit',
+    width: 'inherit',
 });
+
+const filterSidebarStyles = theme =>
+    Object.freeze({
+        searchTab: Object.freeze({
+            '*': {
+                fontFamily: theme.typography.fontFamily,
+            },
+        }),
+    });
 
 class FilterSidebar extends Component {
     componentDidMount() {
@@ -108,20 +116,12 @@ class FilterSidebar extends Component {
             fetchingFeatureFlags,
             embed,
             facilitiesCount,
+            classes,
         } = this.props;
 
         if (fetchingFeatureFlags) {
             return <CircularProgress />;
         }
-
-        const header = !embed ? (
-            <div className="panel-header results-height-subtract">
-                <h3 className="panel-header__title">Open Apparel Registry</h3>
-                <p className="panel-header__subheading">
-                    The open map of global apparel facilities.
-                </p>
-            </div>
-        ) : null;
 
         let orderedTabsForSidebar = vectorTileFeatureIsActive
             ? filterSidebarTabs.slice().reverse()
@@ -179,7 +179,7 @@ class FilterSidebar extends Component {
                             <Tab
                                 key={tab}
                                 label={label}
-                                className="search-tab"
+                                className={`search-tab ${classes.searchTab}`}
                             />
                         );
                     })}
@@ -223,11 +223,7 @@ class FilterSidebar extends Component {
         })();
 
         return (
-            <div
-                className="control-panel"
-                style={filterSidebarStyles.controlPanelStyles}
-            >
-                {header}
+            <div className="control-panel" style={controlPanelStyles}>
                 {tabBar}
                 {insetComponent}
             </div>
@@ -295,4 +291,6 @@ function mapDispatchToProps(dispatch) {
     };
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(FilterSidebar);
+export default withStyles(filterSidebarStyles)(
+    connect(mapStateToProps, mapDispatchToProps)(FilterSidebar),
+);