+++ src/app/src/util/constants.js
@@ -1,6 +1,12 @@
+import includes from 'lodash/includes';
+import { checkWhetherUserHasDashboardAccess } from './util';
+
 export const OTHER = 'Other';
 
 export const FACILITIES_REQUEST_PAGE_SIZE = 50;
+
+export const WEB_HEADER_HEIGHT = '80px';
+export const MOBILE_HEADER_HEIGHT = '68px';
 
 // This choices must be kept in sync with the identical list
 // kept in the Django API's models.py file
@@ -225,7 +231,7 @@ export const facilitiesRoute = '/facilities';
 export const facilityDetailsRoute = '/facilities/:oarID';
 export const claimFacilityRoute = '/facilities/:oarID/claim';
 export const profileRoute = '/profile/:id';
-export const aboutProcessingRoute = '/about/processing';
+export const aboutProcessingRoute = 'https://info.openapparel.org/processing';
 export const dashboardRoute = '/dashboard';
 export const dashboardListsRoute = '/dashboard/lists';
 export const dashboardApiBlocksRoute = '/dashboard/apiblocks';
@@ -243,7 +249,8 @@ export const dashboardGeocoderRoute = '/dashboard/geocoder';
 export const claimedFacilitiesRoute = '/claimed';
 export const claimedFacilitiesDetailRoute = '/claimed/:claimID';
 export const dashboardClaimsDetailsRoute = '/dashboard/claims/:claimID';
-export const aboutClaimedFacilitiesRoute = '/about/claimedfacilities';
+export const aboutClaimedFacilitiesRoute =
+    'https://info.openapparel.org/claimed-facilities';
 
 export const contributeFieldsEnum = Object.freeze({
     name: 'name',
@@ -519,7 +526,7 @@ export const PPE_FIELD_NAMES = Object.freeze([
     'ppe_website',
 ]);
 
-export const OARFont = 'ff-tisa-sans-web-pro,sans-serif';
+export const OARFont = "'DM Sans',sans-serif";
 export const OARColor = '#0427a4';
 
 export const EmbeddedMapInfoLink = 'https://info.openapparel.org/embedded-map';
@@ -527,3 +534,123 @@ export const EmbeddedMapInfoLink = 'https://info.openapparel.org/embedded-map';
 // A CSS size value that is used to set a lower bound on the iframe height
 // when the width is set to 100%
 export const minimum100PercentWidthEmbedHeight = '500px';
+
+export const InfoLink =
+    process.env.NODE_ENV === 'development'
+        ? 'https://oar.niceandserious.com'
+        : 'https://info.openapparel.org';
+
+export const SubmenuLinks = {
+    'How It Works': [
+        {
+            url: 'getting-started',
+            text: 'Getting Started',
+            external: true,
+        },
+        {
+            url: 'data-technology',
+            text: 'Data & Technology',
+            external: true,
+        },
+        { url: 'api', text: 'API', external: true },
+        {
+            url: 'embedded-map',
+            text: 'Embedded Map',
+            external: true,
+        },
+        { url: 'faqs', text: 'FAQs', external: true },
+    ],
+    'About Us': [
+        {
+            url: 'our-mission',
+            text: 'Our Mission',
+            external: true,
+        },
+        {
+            url: 'our-people',
+            text: 'Our People',
+            external: true,
+        },
+        {
+            url: 'governance',
+            text: 'Governance',
+            external: true,
+        },
+        { url: 'funding', text: 'Funding', external: true },
+        { url: 'press', text: 'Press', external: true },
+        {
+            url: 'work-with-us',
+            text: 'Work With Us',
+            external: true,
+        },
+        { url: 'contact', text: 'Contact Us', external: true },
+    ],
+    More: [
+        {
+            text: 'Cookie Preferences',
+            button: true,
+        },
+        {
+            href: `${InfoLink}/privacy-policy`,
+            text: 'Privacy policy',
+            external: true,
+        },
+        {
+            href: `${InfoLink}/terms-of-use`,
+            text: 'Terms of use',
+            external: true,
+        },
+    ],
+};
+
+export const createUserDropdownLinks = (
+    user,
+    logoutAction,
+    activeFeatureFlags,
+) => {
+    const dashboardLink = checkWhetherUserHasDashboardAccess(user)
+        ? Object.freeze([
+              Object.freeze({
+                  text: 'Dashboard',
+                  url: dashboardRoute,
+                  type: 'link',
+              }),
+          ])
+        : [];
+
+    const claimedFacilityLinks = includes(activeFeatureFlags, CLAIM_A_FACILITY)
+        ? Object.freeze([
+              Object.freeze({
+                  text: 'My Facilities',
+                  url: '/claimed',
+                  type: 'link',
+              }),
+          ])
+        : [];
+
+    const userLinks = Object.freeze([
+        Object.freeze({
+            text: 'My Lists',
+            url: '/lists',
+            type: 'link',
+        }),
+        Object.freeze({
+            text: 'Settings',
+            url: '/settings',
+            type: 'link',
+        }),
+    ]);
+
+    const logoutLinks = Object.freeze([
+        Object.freeze({
+            text: 'Log Out',
+            type: 'button',
+            action: logoutAction,
+        }),
+    ]);
+
+    return dashboardLink
+        .concat(claimedFacilityLinks)
+        .concat(userLinks)
+        .concat(logoutLinks);
+};